refactor: Move annotations for reflective class builder to new package
